Add bitwarden-core-macro
