+++ Cargo.toml
@@ -32,6 +32,7 @@ bitwarden-cli = { path = "crates/bitwarden-cli", version = "=2.0.0" }
 bitwarden-collections = { path = "crates/bitwarden-collections", version = "=2.0.0" }
 bitwarden-commercial-vault = { path = "bitwarden_license/bitwarden-commercial-vault", version = "=2.0.0" }
 bitwarden-core = { path = "crates/bitwarden-core", version = "=2.0.0" }
+bitwarden-core-macro = { path = "crates/bitwarden-core-macro", version = "=2.0.0" }
 bitwarden-crypto = { path = "crates/bitwarden-crypto", version = "=2.0.0" }
 bitwarden-encoding = { path = "crates/bitwarden-encoding", version = "=2.0.0" }
 bitwarden-error = { path = "crates/bitwarden-error", version = "=2.0.0" }
