+++ crates/bitwarden-core-macro/src/lib.rs
@@ -1,3 +1,8 @@
+//! Proc macros for the Bitwarden SDK.
+//!
+//! Provides:
+//! - `#[from_client]` attribute macro for generating `from_client` methods on client structs.
+
 use proc_macro::TokenStream;
 use quote::quote;
 use syn::{ItemStruct, parse_macro_input};
@@ -10,7 +15,7 @@ use syn::{ItemStruct, parse_macro_input};
 /// # Example
 ///
 /// ```ignore
-/// use bitwarden_test_macro::from_client;
+/// use bitwarden_core_macro::from_client;
 ///
 /// #[from_client]
 /// pub struct FoldersClient {
@@ -39,6 +44,7 @@ use syn::{ItemStruct, parse_macro_input};
 ///     }
 /// }
 /// ```
+#[proc_macro_attribute]
 pub fn from_client(_attr: TokenStream, item: TokenStream) -> TokenStream {
     let input = parse_macro_input!(item as ItemStruct);
